pyproject.toml
[project]
name = "ailang"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "antlr4-grun>=0.1.0",
    "antlr4-python3-runtime==4.13.2",
    "pandas>=3.0.2",
]
